Reorganize lua directory
