+++ lua/examples/example_of_dynamic_config_request.lua
@@ -1,4 +1,4 @@
-require "dylib"
+require "gatekeeper/dylib"
 
 local acc_start = ""
 local reply_msg = ""
